Moved formatter functions to their own folder.
