+++ src/components/MovieDetails/index.jsx
@@ -3,16 +3,7 @@ import { Poster } from "../Poster";
 import { MovieGenres } from '../MovieGenres';
 
 import './MovieDetails.css';
-
-function formatDuration(durationInMinutes) {
-    if (isNaN(durationInMinutes)) {
-        return 'N/A';
-    }
-    const hours = Math.floor(durationInMinutes / 60);
-    const minutes = durationInMinutes - (hours * 60);
-
-    return `${hours > 0 ? `${hours}h ` : ''}${minutes}min`;
-}
+import { formatDuration } from '../../formatters';
 
 function MovieDetails({ movieData }) {
     const {
