+++ src/formatters/index.js
@@ -1,0 +1,4 @@
+import { genresText } from "./genresText";
+import { formatDuration } from "./formatDuration";  
+
+export { genresText,  formatDuration }
